+++ compiler-cli/src/new.rs
@@ -14,7 +14,7 @@ mod tests;
 
 use crate::NewOptions;
 
-const GLEAM_STDLIB_VERSION: &str = "0.30";
+const GLEAM_STDLIB_VERSION: &str = "0.31";
 const GLEEUNIT_VERSION: &str = "0.10";
 const ERLANG_OTP_VERSION: &str = "26.0.2";
 const REBAR3_VERSION: &str = "3";
@@ -176,8 +176,8 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3.5.3
-      - uses: erlef/setup-beam@v1.16.0
+      - uses: actions/checkout@v3
+      - uses: erlef/setup-beam@v1
         with:
           otp-version: "{}"
           gleam-version: "{}"